requirements.txt
pandas==2.2.3
pyarrow==18.1.0
numpy
matplotlib==3.9.2
#huggingface_hub==0.26.2
transformers==4.46.2
requests==2.32.3
#sentencepiece==0.2.0
#langchain==0.3.7
#langchain_community==0.3.7
google-generativeai==0.8.3
#langchain-google-genai==2.0.4
streamlit==1.40.1
#tiktoken==0.8.0
bs4
plotly==5.24.1
streamlit_extras==0.5.0
#scikit-learn==1.5.2
#nltk==3.9.1
#rouge-score==0.1.2
#kaleido==0.2.1
#nbformat==5.10.4
seaborn==0.13.2